Do not set the color if it is the same as the cluster color
The value will be set the second time an ink gets saved as the getter
for color takes it out of the cluster_color field if color is not present.
That then means that values that existing values get solidified and it
is not really possible to change it much later on.

This fixes it for new inks. The old data will have to be migrated to
remove duplicates though.