Replace webpacker with shakapacker
Webpacker is retired. At the moment the recommendations from Rails
seems to be either Import maps (which is still a fairly cutting edge
feature requiring a shim for Safari and HTTP/2 all the way) or
jsbundling-rails. Sadly, jsbundling-rails does not have support for
HMR. They recommend shakapacker if that feature is important, which
I think it is.

Fixes #1358