Remove ability to delete pens and inks
The distinction between archiving and deleting an ink is super confusing
to anyone who isn't a programmer, it seems. Mainly, because the deletion
only works when the pen or ink hasn't been used for a currently inked
entry. Now, you can only archive pens and inks and be content with that. ;)

Fixes #1381