Use same ink color cell as collected inks
Bootstrap apparently uses box shadow to fill in the colors of the striped
tables. This meant any background-color set on the table is overlaid with
the box-shadow color. This was a bit unexpected. The collected inks table
was unaffected by this since it set the background color on a div inside
the cell, rather than the cell itself. Let's do the same in these tables
as well. Bonus: with the same ink color size as collected inks, there's
now no longer a layout shift when the Add to collection buttons load.

Fixes #1425