Sort currently inked entries by pen name by default
We do this sorting in the table, but not in the card view so when
switching from the desktop to a mobile phone "stuff changes".

At some point I imagine we will replace this by a way to sort entries in
the card view, but until then let's just make it the same as the default
in the table view.