Separate the leaderboard refresh workers for better stats
Some (one?) of these refresh workers (at least the "brands" one) puts
a lot of strain on the database. It's hard to get any stats on which one
is problematic when they're all using the same worker. This separates
them all so that we get separate run times that are easier to interpret.