Fix add to collection button when many results
When there are many results to a search, we would make a request to the backend
for each of them. That could mean, that we are running into the rate limiting
and then some of them would randomly not work.

So instead, let's just make requests for those that are on screen and also
increase the rate limit somewhat.