Layer between finding pages/videos to review and the actual reviewing
That way the job of doing the finding of things to review and that of actually
doing the matching is separated and each part will be shorter.

Also, we are now keeping track of what pages have already been processed in the
database, so now the Redis cache is really just a cache and not the source of
truth anymore and can just be cleared.