Review ink clusterer runs
Now that the ink clusterer runs on every new micro cluster that is not attached
to anything, we can change this controller to only do the review part.

This sets the stage for also doing reviews later one after the fact (i.e. when
we just let the follow up agents to the actions and just review and undo when
necessary).